Add files via upload
This file contains the data set of some job posting. The data file contains the following attributes : job_id, title, location,	department, salary_range, company_profile,	description, requirements, benefits, telecommuting, has_company_logo, has_questions, employment_type, required_experience, required_education,	industry,	function and 	fraudulent. The fraudulent indicates whether the posting is fake or real. If the value is "0" it is a real post. If the value is "1" it is a fake post.